pyproject.toml
[tool.poetry]
name = "azcmd"
version = "0.1.0"
description = "Simplify dealing with Azure Blob Storages."
authors = ["Jason Viloria <jnvilo@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
azure-identity = "^1.15.0"
azure-storage-blob = "^12.18.3"
azure-mgmt-resource = "^23.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
az-get = "azcmd.azget:main"
azblob = "azcmd.azblob:main"
